Use the ‘&’ constraint modifier in inline assembly
Use the ‘&’ constraint modifier (see Constraint Modifier Characters) on all output operands that must not overlap an input. Otherwise, GCC may allocate the output operand in the same register as an unrelated input operand, on the assumption that the assembler code consumes its inputs before
producing outputs. This assumption may be false if the assembler code actually consists of more than one instruction.

Ref: https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html